Movie CRUD Operations

+++ server/Controller/MovieController.js
@@ -1,5 +1,6 @@
 import Movie from "../Schema/movie.js";
 
+//Add Movie
 export const AddMovie = async (req, res) => {
   if (req.user.isAdmin) {
     try {
@@ -15,24 +16,81 @@ export const AddMovie = async (req, res) => {
   }
 };
 
+//Update Movie
+
 export const UpdateMovie = async (req, res) => {
+  if (req.user.isAdmin) {
+    try {
+      const updatedmovie = await Movie.findByIdAndUpdate(
+        req.params.id,
+        { $set: req.body },
+        { new: true }
+      );
+      res.status(200).json(updatedmovie);
+    } catch (err) {
+      res.status(400).json(err);
+    }
+  } else {
+    res.status(400).json("Only admin allowed");
+  }
+};
+
+//Delete Movie
+export const DeleteMovie = async (req, res) => {
+  if (req.user.isAdmin) {
+    try {
+      await Movie.findByIdAndDelete(req.params.id);
+      res.status(200).json("Movie Deleted");
+    } catch (err) {
+      res.status(400).json(err);
+    }
+  } else {
+    res.status(400).json("Only admin allowed");
+  }
+};
+
+//Get Single Movie
+export const SingleMovie = async (req, res) => {
   try {
-    const updatedmovie = await Movie.findByIdAndUpdate(
-      req.params.id,
-      { $set: req.body },
-      { new: true }
-    );
-    res.status(200).json(updatedmovie);
+    const movie = await Movie.findById(req.params.id);
+    res.status(200).json(movie);
   } catch (err) {
     res.status(400).json(err);
   }
 };
 
-export const DeleteMovie = async (req, res) => {
+//Get random Movie
+
+export const RandomMovie = async (req, res) => {
+  const type = req.query.type;
+  let movie;
   try {
-    await Movie.findByIdAndDelete(req.params.id);
-    res.status(200).json("Movie Deleted");
+    if (type === "series") {
+      movie = await Movie.aggregate([
+        { $match: { isSeries: true } },
+        { $sample: { size: 1 } },
+      ]);
+    } else {
+      movie = await Movie.aggregate([
+        { $match: { isSeries: false } },
+        { $sample: { size: 1 } },
+      ]);
+    }
+    res.status(200).json(movie);
   } catch (err) {
-    res.status(400).json(err);
+    res.send(400).json(err);
+  }
+};
+
+//All Movie
+
+export const AllMovies = async (req, res) => {
+  if (req.user.isAdmin) {
+    try {
+      const movies = await Movie.find();
+      res.status(200).json(movies);
+    } catch (err) {
+      res.status(400).json(err);
+    }
   }
 };
